useTableValues hook.
