+++ src/features/vote/styled/CommitPhase.styled.tsx
@@ -22,6 +22,14 @@ export const FormWrapper = styled.form<StyledFormProps>`
         }
       }
     }
+    .status {
+      p {
+        margin: 0 auto;
+      }
+    }
+    .center-header {
+      text-align: center;
+    }
     .table {
       ${tw`table-auto`};
       width: 100%;
@@ -55,7 +63,6 @@ export const FormWrapper = styled.form<StyledFormProps>`
         td {
           div {
             display: flex;
-            align-items: center;
           }
           .description {
             max-width: 500px;
